+++ WooCommerce/src/main/kotlin/com/woocommerce/android/analytics/AnalyticsTracker.kt
@@ -198,6 +198,7 @@ class AnalyticsTracker private constructor(private val context: Context) {
         const val KEY_REASON = "reason"
         const val KEY_TAP = "tap"
         const val KEY_FAILURE = "failure"
+        const val KEY_IS_FREE_TRIAL = "is_free_trial"
 
         const val KEY_SORT_ORDER = "order"
         const val VALUE_SORT_NAME_ASC = "name,ascending"
@@ -316,6 +317,10 @@ class AnalyticsTracker private constructor(private val context: Context) {
 
         // -- Linked Products
         const val KEY_LINKED_PRODUCTS_ACTION = "action"
+
+        // -- Product Selector
+        const val KEY_PRODUCT_SELECTOR_SOURCE = "source"
+        const val KEY_PRODUCT_SELECTOR_FILTER_STATUS = "is_filter_active"
 
         enum class LinkedProductsAction(val value: String) {
             SHOWN("shown"),
@@ -490,7 +495,6 @@ class AnalyticsTracker private constructor(private val context: Context) {
 
         // -- Domain change
         const val VALUE_SETTINGS = "settings"
-        const val VALUE_ONBOARDING = "onboarding"
         const val VALUE_STEP_DASHBOARD = "dashboard"
         const val VALUE_STEP_PICKER = "picker"
         const val VALUE_STEP_CONTACT_INFO = "contact_info"
@@ -508,6 +512,10 @@ class AnalyticsTracker private constructor(private val context: Context) {
         const val VALUE_ADD_DOMAIN = "add_domain"
         const val VALUE_LAUNCH_SITE = "launch_site"
         const val VALUE_PAYMENTS = "payments"
+
+        // -- Product Selector
+        const val VALUE_PRODUCT_SELECTOR = "product_selector"
+        const val VALUE_VARIATION_SELECTOR = "variation_selector"
 
         var sendUsageStats: Boolean = true
             set(value) {