+++ WooCommerce/src/main/kotlin/com/woocommerce/android/ui/login/jetpack/wpcom/JetpackActivationWPComEmailViewModel.kt
@@ -53,6 +53,7 @@ class JetpackActivationWPComEmailViewModel @Inject constructor(
     }
 
     fun onCloseClick() {
+        wpComLoginRepository.clearAccessToken()
         triggerEvent(Exit)
 
         analyticsTrackerWrapper.track(