+++ WooCommerce/src/test/kotlin/com/woocommerce/android/ui/products/ProductDetailViewModel_AddFlowTest.kt
@@ -176,7 +176,8 @@ class ProductDetailViewModel_AddFlowTest : BaseUnitTest() {
                 generateVariationCandidates,
                 mock(),
                 tracker,
-                selectedSite
+                selectedSite,
+                mock()
             )
         )